pyproject.toml
[tool.poetry]
name = "kubernetes-ext"
version = "0.0.1"
description = "`kubernetes-ext` is a Meltano utility extension."
authors = ["Josh Bielick"]
license = "Apache 2.0"
packages = [
    { include = "kubernetes_ext" }
    # uncomment this if you want to include static assets
    # { include = "files_kubernetes-ext.*" }
]
# If you need to static assets with your extension, you can add them here.
# Under a directory path of "files_kubernetes-ext/"
#include = [
#    "files_kubernetes-ext/somefile.thing"
#]

[tool.poetry.dependencies]
python = "<4,>=3.8"
PyYAML = "^6.0.0"
click = "^8.1.3"
typer = "^0.9.0"
"meltano.edk"= {git = "https://github.com/meltano/edk.git", rev="v0.4.0"}
jinja2 = "^3.1.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kubernetes = 'kubernetes_ext.main:app'
